Intializing the appointement module and Major fixes after testing ten…

+++ Modules/HotelBooking/Database/Migrations/2023_11_29_114048_rename_city_id_to_state_id_in_booking_informations_table.php
@@ -13,11 +13,13 @@ class RenameCityIdToStateIdInBookingInformationsTable extends Migration
      */
     public function up()
     {
-        Schema::table('booking_informations', function (Blueprint $table) {
-            $table->dropForeign(['city_id']);
-            $table->renameColumn('city_id', 'state_id');
-            $table->foreign('state_id')->references('id')->on('states')->onDelete('cascade');
-        });
+        if (Schema::hasTable('booking_informations') && !Schema::hasColumn('booking_informations', 'city_id')) {
+            Schema::table('booking_informations', function (Blueprint $table) {
+                $table->dropForeign(['city_id']);
+                $table->renameColumn('city_id', 'state_id');
+                $table->foreign('state_id')->references('id')->on('states')->onDelete('cascade');
+            });
+        }
     }
 
     /**
